chore: address issue #146
